+++ lib/PhpParser/PrettyPrinter/Standard.php
@@ -23,7 +23,8 @@ class Standard extends PrettyPrinterAbstract {
              . ($node->byRef ? '&' : '')
              . ($node->variadic ? '...' : '')
              . $this->p($node->var)
-             . ($node->default ? ' = ' . $this->p($node->default) : '');
+             . ($node->default ? ' = ' . $this->p($node->default) : '')
+             . ($node->hooks ? ' {' . $this->pStmts($node->hooks) . $this->nl . '}' : '');
     }
 
     protected function pArg(Node\Arg $node): string {
@@ -123,6 +124,10 @@ class Standard extends PrettyPrinterAbstract {
 
     protected function pScalar_MagicConst_Trait(MagicConst\Trait_ $node): string {
         return '__TRAIT__';
+    }
+
+    protected function pScalar_MagicConst_Property(MagicConst\Property $node): string {
+        return '__PROPERTY__';
     }
 
     // Scalars
@@ -827,12 +832,22 @@ class Standard extends PrettyPrinterAbstract {
         return $this->pAttrGroups($node->attrGroups)
             . (0 === $node->flags ? 'var ' : $this->pModifiers($node->flags))
             . ($node->type ? $this->p($node->type) . ' ' : '')
-            . $this->pCommaSeparated($node->props) . ';';
+            . $this->pCommaSeparated($node->props)
+            . ($node->hooks ? ' {' . $this->pStmts($node->hooks) . $this->nl . '}' : ';');
     }
 
     protected function pPropertyItem(Node\PropertyItem $node): string {
         return '$' . $node->name
              . (null !== $node->default ? ' = ' . $this->p($node->default) : '');
+    }
+
+    protected function pPropertyHook(Node\PropertyHook $node): string {
+        return $this->pAttrGroups($node->attrGroups)
+             . $this->pModifiers($node->flags)
+             . ($node->byRef ? '&' : '') . $node->name
+             . ($node->params ? '(' . $this->pMaybeMultiline($node->params, $this->phpVersion->supportsTrailingCommaInParamList()) . ')' : '')
+             . (\is_array($node->body) ? ' {' . $this->pStmts($node->body) . $this->nl . '}'
+                : ($node->body !== null ? ' => ' . $this->p($node->body) : '') . ';');
     }
 
     protected function pStmt_ClassMethod(Stmt\ClassMethod $node): string {