+++ test/fixtures/Suit.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+enum Suit {
+    case Hearts;
+    case Diamonds;
+    case Clubs;
+    case Spades;
+}